A fix to the getters.
The getters are ultimately returning an ArrayList<String> so they should be set to that variable type. I don't know if that gave you an error on your end, but setting them to ArrayList<String> fixed the checked exception and satisfied the logical discrepancy I saw in it. Further input on the matter is appreciated. Everything works great though. Great job on the usage of the loop and StringBuilder.